Fix #8 Authentication Broken
Implementation was not correct and its test was returning a false
positive.

A couchdb db with no members in it is public. The test was lacking
adding a member to the database.

The implementation was performing the authentication using the URI::HTTP 'build' method
In its documentation we can read information about the problem:
http://ruby-doc.org/stdlib-2.0.0/libdoc/uri/rdoc/URI/HTTP.html#method-c-build-label-Description
 ```
   Currently, if passed userinfo components this method generates
   invalid HTTP URIs as per RFC 1738.
   ```

This PR uses this procedure:
http://ruby-doc.org/stdlib-2.2.1/libdoc/net/http/rdoc/Net/HTTP.html#class-Net::HTTP-label-Basic+Authentication